Move TimelineItemContent and related types into a new submodule
